replace Queryset protocol by abstract base type
